+++ sourcespec/ssp_summary_statistics.py
@@ -172,20 +172,22 @@ def _param_summary_statistics(
     return summary
 
 
-def _make_summary_spec(station, channel, freq_logspaced, Mw, fc, t_star):
+def _make_summary_spec(station, channel, freq_logspaced, Mw, fc, t_star,
+                       config):
     """Create a synthetic spectrum for summary statistics."""
     sp = Spectrum()
     sp.stats.station = station
     sp.stats.channel = channel
     sp.freq_logspaced = freq_logspaced
-    data_mag_logspaced = spectral_model(freq_logspaced, Mw, fc, t_star)
+    data_mag_logspaced = spectral_model(
+        freq_logspaced, Mw, fc, t_star, n=config.source_falloff_power)
     # data_logspaced must be provided before data_mag_logspaced
     sp.data_logspaced = mag_to_moment(data_mag_logspaced)
     sp.data_mag_logspaced = data_mag_logspaced
     return sp
 
 
-def _add_summary_spectra(sspec_output, spec_st):
+def _add_summary_spectra(sspec_output, spec_st, config):
     """Add to the spectra stream synthetic spectra for summary statistics."""
     fmins = [np.nanmin(spec.freq) for spec in spec_st]
     fmaxs = [np.nanmax(spec.freq) for spec in spec_st]
@@ -204,13 +206,15 @@ def _add_summary_spectra(sspec_output, spec_st):
             t_star = t_star_model
             break
     spec_st.append(
-        _make_summary_spec('SUMMARY', 'SSS', freq_logspaced, Mw, fc, t_star)
+        _make_summary_spec(
+            'SUMMARY', 'SSS', freq_logspaced, Mw, fc, t_star, config)
     )
     spec_st.append(
-        _make_summary_spec('SUMMARY', 'SSs', freq_logspaced, Mw, fc, 0)
+        _make_summary_spec('SUMMARY', 'SSs', freq_logspaced, Mw, fc, 0, config)
     )
     spec_st.append(
-        _make_summary_spec('SUMMARY', 'SSt', freq_logspaced, Mw, 1e999, t_star)
+        _make_summary_spec(
+            'SUMMARY', 'SSt', freq_logspaced, Mw, 1e999, t_star, config)
     )
 
 
@@ -390,7 +394,7 @@ def compute_summary_statistics(config, sspec_output, spec_st, weight_st):
     logger.info(f'params_percentiles: {sourcepar_percentiles}')
 
     # Add synthetic spectra for summary statistics
-    _add_summary_spectra(sspec_output, spec_st)
+    _add_summary_spectra(sspec_output, spec_st, config)
     # Compute dispersion around the summary synthetic spectrum
     logger.info(
         'Computing dispersion around the summary synthetic spectrum...')